requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
supabase==2.0.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx>=0.24.0,<0.25.0
python-multipart>=0.0.7
email-validator==2.1.0
